requirements.txt
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.13.4
binaryornot==0.4.4
bleach==6.2.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
cryptography==44.0.3
Django==5.2
django-browser-reload==1.18.0
django-environ==0.12.0
django-markdownify==0.9.5
django-ninja==1.4.1
django-recaptcha==4.1.0
django-taggit==6.1.0
django-tailwind==4.0.1
django-unfold==1.0.0
gunicorn==23.0.0
h11==0.16.0
html2text==2025.4.15
idna==3.10
Jinja2==3.1.6
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
pillow==11.2.1
psycopg==3.2.7
psycopg-binary==3.2.7
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-http-client==3.3.7
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
sendgrid==6.12.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
text-unidecode==1.3
tinycss2==1.4.0
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
webencodings==0.5.1
Werkzeug==3.1.3
